refactoring parse -> merge
